+++ app/.storybook/i18next.js
@@ -3,7 +3,7 @@ import LanguageDetector from "i18next-browser-languagedetector";
 // Configure i18next for storybook addon storybook-react-i18next
 // See https://storybook.js.org/addons/storybook-react-i18next
 import i18n from "i18next";
-import i18nConfig from "../next-i18next.config";
+import i18nConfig from "../i18n.config";
 import { initReactI18next } from "react-i18next";
 
 const ns = ["common"];
