+++ app/i18n.config.js
