+++ app/next.config.js
@@ -1,6 +1,6 @@
 /** @type {import('next').NextConfig} */
 
-const { i18n } = require("./next-i18next.config");
+const { i18n } = require("./i18n.config");
 
 const nextConfig = {
   i18n,
